Authorization-to-complete

+++ routes/reportRoute.js
@@ -1,29 +1,30 @@
 const express = require('express');
 const mongoose = require('mongoose');
 const router = express.Router();
-const isLoggedIn = require('../middleware');
-const isOwner = require('../middleware');
 const asyncHandler = require('express-async-handler');
 
 const ExpressError = require('../utils/ExpressError')
 const Reports = require('../model/reportModel');
 
+// const validateToken = require('../middleware/validatorMiddleware');
+const validateToken = require('../middlewares/validatorMiddleware');
+
 // Get All Reports 
 // Route /reports
-router.get('/reports', isLoggedIn, async(req, res) => {
-    let reports = await Reports.find({});
+router.get('/reports', validateToken, async(req, res) => {
+    let reports = await Reports.find({user_id: req.userAvailable.id});
     res.render("UI/reports.ejs", {reports});
 });
 
 // Get create report
 // Route /newreport
-router.get('/newreport', isLoggedIn, (req, res) => {
+router.get('/newreport', validateToken, (req, res) => {
     res.render('UI/newreport.ejs');
 });
 
 // Get particular report
 // Route /reports/:id
-router.get('/reports/:id', isLoggedIn, isOwner, asyncHandler(async(req, res) => {
+router.get('/reports/:id', validateToken, asyncHandler(async(req, res) => {
     const {id} = req.params;
     if (!mongoose.Types.ObjectId.isValid(id)) {
         throw new ExpressError(404, "Id not found");
@@ -38,25 +39,33 @@ router.get('/reports/:id', isLoggedIn, isOwner, asyncHandler(async(req, res) => 
 
 // Post create report
 // Route /reports
-router.post('/reports', isLoggedIn, isOwner, async (req, res) => {
+router.post('/reports', validateToken, async (req, res) => {
     let { name, incomeAmount, description} = req.body;
+    // validation
+    if(!name || !incomeAmount || !description) {
+        throw new ExpressError(400, "All Fields Mandatory");
+    }
     // creating new Report in db
     let newReport = new Reports({
+        user_id: req.userAvailable.id,
         name: name,
         incomeAmount: incomeAmount,
         description: description
     }); 
-    // console.log(newReport);
     await newReport.save();
+    // req.flash("success", "New Report Created");
     res.redirect('/reports');
 });
 
 // Delete a Report
 // /reports/:id
-router.delete('/reports/:id', isLoggedIn, isOwner, asyncHandler(async(req, res) => {
+router.delete('/reports/:id', validateToken, asyncHandler(async(req, res) => {
     let {id} = req.params;
     if (!mongoose.Types.ObjectId.isValid(id)) {
         throw new ExpressError(404, "Id not found");
+    }
+    if(Reports.user_id.toString() !== req.userAvailable.id) {
+        throw new ExpressError(403, "User dont have permission to update other user contact");
     }
     await Reports.findByIdAndDelete(id);
     res.redirect('/reports');
